Update material_and_questions.md
The exclusion of all negative answers from MA question types, I think, is an overreaction to the fact that sometimes a negative answer does not answer the question and should not be valid. Even the second example given in the rulebook (which I propose deleting) should be a valid MA question - "A crown that will not last" is a type of crown defined by its temporary character even as "a crown that will last forever" is defined by its eternal character. These are two different crowns. Also, this rule also makes Romans 8:38-39, one of the great Scripture verses about the love and power of God, unuseable as an MA ("What will be able to separate us from the love of God that is in Christ Jesus our Lord?"), which is the only way it could make it into the question set other than as a Quote or FTV. Again, I think this is a rule that overreaches in that it is attempting to correct a situation for some verses that someone might pull as an MA by eliminating all - even good - negative answers from the question pool of MAs. Whether a negative answer is valid or invalid needs to be reviewed on a case-by-case basis, meaning, is the answer to the question actually two or more different nouns, adjectives, or adverbs that would answer the interrogative.